updated format for SegmentSpeedMap.  Adds conversion to kph  (#46)
* added logic to allow for mph or kph in the data.

* added tests for records with units.

* format

* script to fix formatting

* 0.2.2-dev [publish binary]

* unint32_t way too big.

* added limit to speeds.  throw exception if > invalid_speed - 1.

* added max tests.

* 0.2.2-dev [publish binary]

* 0.2.2-dev [publish binary]

* 0.2.2-dev [publish binary]

* 0.2.2-dev.1 [publish binary]

* updated.

* updated version.

* allow importing of data even if speed data is bad.

* format updates.

* v0.2.2-dev.2 [publish binary]

* added mph and kph tests.

* added logic to convert mph to kph for maxspeed tag.

* v0.2.2-dev.3 [publish binary]

* updated.

+++ src/types.hpp
@@ -10,7 +10,7 @@
 typedef std::uint64_t external_nodeid_t;
 typedef std::uint32_t internal_nodeid_t;
 typedef std::uint32_t wayid_t;
-typedef std::uint32_t segment_speed_t;
+typedef std::uint8_t segment_speed_t;
 
 // Way ID, and whether it the node pair for it is stored forward or backward
 typedef struct
